Fix LCD servo voltage and first measurement state
Changed calculation for servo voltage based on the voltage divider
used. A division between integers was resulting in an integer instead
of a float, changing the expected value.

Also fixes the initial measurement state by inverting the digitalRead()
returned value. In previous versions of the program, it was assumed the
button would be pulled up by a resistor, when in the board it actually
is pulled down.

Closes #2 and #3