Add manual calibration and change default sample values
Added manual calibration mode by holding the button for 2 seconds (default).
This is achieved by changing the hold variable in main.ino to change only
when the button press is released. A manualCalibration() function was created
in sampling.cpp to handle the task.

Also changed real time sampling value to 500 (was 200) and average sampling
value to 10000 (was 5000). The first was important because the numbers were
changing very quickly on the display, making it really hard to see sudden changes.

Deleted an unused variable in main.ino, 'mamp', which is now in calibration.cpp.

Closes #4